Initial Commit: Convert to node module with npm init.
