+++ .babelrc
@@ -1,0 +1,4 @@
+{
+  "stage": 0,
+  "loose": "all"
+}
