--- hasher.nmf
@@ -1,9 +1,0 @@
-{
-  "program": {
-    "portable": {
-      "pnacl-translate": {
-        "url": "hasher.pexe"
-      }
-    }
-  }
-}
